fix: timer types

+++ AresOS/Kernel/include/drivers/time.h
@@ -1,9 +1,11 @@
 #ifndef _TIME_H_
 #define _TIME_H_
 
+#include <stdint.h>
+
 void timer_handler();
-int ticks_elapsed();
-int seconds_elapsed();
+uint64_t ticks_elapsed();
+uint64_t seconds_elapsed();
 void sleep(int seconds);
 
 #endif
